Polish scenario test, doc
